skill stream routes
