+++ src/routes/paths.ts
@@ -79,7 +79,8 @@ export const coLinksPaths = {
   exploreSkill: (skill: string) => `/explore/interests/${skill}`,
   exploreSkills: `/explore/interests`,
   give: '/give',
-  giveSkill: (skill: string) => `/give/leaderboard/${skill}`,
+  giveSkill: (skill: string) => `/give/skill/${skill}`,
+  giveSkillLeaderboard: (skill: string) => `/give/leaderboard/${skill}`,
   giveSkillMap: (skill: string) => `/give/map/${skill}`,
   deprecated_giveParty: `/giveparty`,
   deprecated_givePartyAddress: (address: string) => `/giveparty/${address}`,
